Fresh start on keyboard layout context module
